backend/voice-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-engine"
version = "0.1.0"
description = "Wanda Voice Engine (Orb + MCC) - offline-first"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-ulid>=2.7.0",
  "httpx>=0.27",
  "numpy>=1.26",
  "sounddevice>=0.4.6",
  "faster-whisper>=1.0.3",
  "torch>=2.2.0",
  "f5-tts>=0.1.0",
]

[project.scripts]
voice-engine = "voice_engine.cli:main"
